+++ src/main/java/co/invest72/investment/domain/InvestmentAmount.java
@@ -3,7 +3,7 @@ package co.invest72.investment.domain;
 import java.math.BigDecimal;
 
 public interface InvestmentAmount {
-	double calAnnualInterest(InterestRate interestRate);
+	BigDecimal calAnnualInterest(InterestRate interestRate);
 
 	BigDecimal calMonthlyInterest(InterestRate interestRate);
