+++ src/main/java/co/invest72/investment/presentation/response/CalculateInvestmentResponse.java
@@ -1,5 +1,6 @@
 package co.invest72.investment.presentation.response;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Builder;
@@ -9,18 +10,18 @@ import lombok.Getter;
 @EqualsAndHashCode
 @Getter
 public class CalculateInvestmentResponse {
-	private final int totalInvestment;
-	private final int totalInterest;
-	private final int totalTax;
-	private final int totalProfit;
+	private final BigDecimal totalInvestment;
+	private final BigDecimal totalInterest;
+	private final BigDecimal totalTax;
+	private final BigDecimal totalProfit;
 	private final String taxType;
 	private final String taxPercent;
 	private final List<MonthlyInvestmentResult> monthlyDetails;
 	private final List<YearlyInvestmentResult> yearlyDetails;
 
 	@Builder
-	public CalculateInvestmentResponse(int totalInvestment, int totalInterest, int totalTax,
-		int totalProfit, String taxType, String taxPercent, List<MonthlyInvestmentResult> monthlyDetails,
+	public CalculateInvestmentResponse(BigDecimal totalInvestment, BigDecimal totalInterest, BigDecimal totalTax,
+		BigDecimal totalProfit, String taxType, String taxPercent, List<MonthlyInvestmentResult> monthlyDetails,
 		List<YearlyInvestmentResult> yearlyDetails) {
 		this.totalInvestment = totalInvestment;
 		this.totalInterest = totalInterest;
